scripts/requirements.txt
sentence-transformers==2.5.1
python-dotenv==1.0.1
aiohttp==3.9.3
tqdm==4.66.2
pinecone-client==3.0.1 